refactor: shared useDeferredSearch hook for list components
